# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online image shop API",
        "description": "API za upravljanje porudžbinama, slikama, tehnikama slikanja, popustima i korisnicima kroz autentifikaciju i autorizaciju",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/slike": {
            "get": {
                "tags": [
                    "Slike"
                ],
                "summary": "Vraća sve slike",
                "operationId": "e819b402d6525e8e5db0ce16239ca1dc",
                "responses": {
                    "200": {
                        "description": "Lista slika"
                    }
                }
            },
            "post": {
                "tags": [
                    "Slike"
                ],
                "summary": "Dodavanje nove slike",
                "operationId": "582832efe61d7ba00bfd7dc770880bf3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "cena",
                                    "visina_cm",
                                    "sirina_cm",
                                    "dostupna",
                                    "tehnike"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Sunset"
                                    },
                                    "cena": {
                                        "type": "number",
                                        "example": 15000
                                    },
                                    "visina_cm": {
                                        "type": "number",
                                        "example": 50
                                    },
                                    "sirina_cm": {
                                        "type": "number",
                                        "example": 70
                                    },
                                    "dostupna": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "galerija_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "putanja_fotografije": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "tehnike": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slika kreirana"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/slike/filter": {
            "get": {
                "tags": [
                    "Slike"
                ],
                "summary": "Vraća paginirane i filtrirane slike",
                "operationId": "73f73818d7a25196ccfddf583443efaa",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 20,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "dostupna",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "cena_min",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "cena_max",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "visina_cm_min",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "visina_cm_max",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "sirina_cm_min",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "sirina_cm_max",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "sort_cena",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "sort_starost",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "tehnike[]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginirana lista slika"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/slike/{id}": {
            "get": {
                "tags": [
                    "Slike"
                ],
                "summary": "Vraća jednu sliku",
                "operationId": "91694b9773fbd6702535c7158b843051",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slika"
                    },
                    "404": {
                        "description": "Slika nije pronađena"
                    }
                }
            },
            "post": {
                "tags": [
                    "Slike"
                ],
                "summary": "Ažurira sliku (koristi POST + _method=PUT zbog multipart/form-data)",
                "operationId": "d1fffc73b647e40a70d39925a24da75f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "galerija_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "putanja_fotografije": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cena": {
                                        "type": "number",
                                        "example": 200
                                    },
                                    "naziv": {
                                        "type": "string",
                                        "example": "Jutarnji pejzaž"
                                    },
                                    "visina_cm": {
                                        "type": "number",
                                        "example": 60
                                    },
                                    "sirina_cm": {
                                        "type": "number",
                                        "example": 80
                                    },
                                    "dostupna": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tehnike": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slika ažurirana"
                    },
                    "400": {
                        "description": "Nema podataka za izmenu"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Slike"
                ],
                "summary": "Briše sliku",
                "operationId": "3352072bc66871fd068f669fdfa579f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slika obrisana"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Slika nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tehnike": {
            "get": {
                "tags": [
                    "Tehnike"
                ],
                "summary": "Vraća sve tehnike",
                "operationId": "c21a471aae53ba8124b6b418673bb77f",
                "responses": {
                    "200": {
                        "description": "Lista tehnika"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tehnike"
                ],
                "summary": "Kreira novu tehniku",
                "operationId": "1a8aa5e0eb95f088b18408d35f1a150b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Akvarel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tehnika kreirana"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}